apps/api/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-bcrypt
PyJWT
python-dotenv
PyPDF2
PyMuPDF
llama-index
openai
gunicorn
pytest
requests
